% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_rmcab_data.R
\name{download_rmcab_data}
\alias{download_rmcab_data}
\title{Dowload air quality and meterological data from air quality station of RMCAB}
\usage{
download_rmcab_data(aqs_code, start_date, end_date)
}
\arguments{
\item{aqs_code}{Integer with the code of the station. See rmcab_aqs().}

\item{start_date}{Date to start downloading in dd-mm-yyyy.}

\item{end_date}{Date to start downloading in dd-mm-yyyy.}
}
\value{
data frame with station data.
}
\description{
This function download the data from one station of Bogota air quality
network. Missing data is in NA.
}
\examples{
\dontrun{
# Download data from Usaquen from first week of February
usaquen_code <- 1 # rmcab_aqs
usaquen <- download_rmcab_data(usquen_code, "01-02-2023", "07-02-2023")
}
}
