% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rmcab_aqs}
\alias{rmcab_aqs}
\title{RMCAB AQS station latitude and longitude}
\format{
A data frame with 20 observations and 4 variables:
\describe{
\item{aqs}{RMCAB AQS name.}
\item{code}{RMCAB AQS code.}
\item{lat}{RMCAB AQS latitude.}
\item{lon}{RMCAB AQS longitude.}
}
}
\usage{
rmcab_aqs
}
\description{
RMCAB air quality stations (AQS) latitudes and longitudes.
Use this to check the AQS \code{aqs_code} argument in
\code{download_rmcab_data()}.
#'
}
\examples{
rmcab_aqs
}
\keyword{datasets}
