% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rmcab_params}
\alias{rmcab_params}
\title{RMCAB AQS parameters}
\format{
A data frame with 13 observations and 2 variables:
\describe{
\item{name}{RMCAB parameter name.}
\item{units}{Parameter unit.}
}
}
\usage{
rmcab_params
}
\description{
RMCAB air quality stations parameters
Use this to check the units of columsn in returned data.frame
#'
}
\examples{
rmcab_params
}
\keyword{datasets}
